add float

+++ train_ann.py
@@ -32,7 +32,7 @@ from timebase.utils.scheduler import Scheduler
 
 def load(d: t.Dict[str, torch.Tensor], device: torch.device):
     """Load values in dictionary d to device"""
-    return {k: v.to(device) for k, v in d.items()}
+    return {k: v.float().to(device) for k, v in d.items()}
 
 
 def load_pre_trained_parameters(
